refactor(server): split Indexer into IndexQuery and BackgroundIndexer
Separate the read-only query surface (IndexQuery) from the background
indexing scheduler (BackgroundIndexer), introduce a SessionStore that
owns the open-document table and buffer sync, and report indexing
progress through a signal instead of a borrowed transport peer. Pure
code movement with no behavior change.
